pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jax-prng-repro"
version = "0.1.0"
description = "Reproducibility playbook for JAX: PRNG keys, fold_in, checkpoint-safe RNG, deterministic hashes."
authors = [{name="Mohsin's PRNG Repro", email="example@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jax>=0.4.26",
    "flax>=0.7",
    "optax>=0.2.2",
    "orbax-checkpoint>=0.6.1",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "ruff>=0.5.0",
    "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E","F","I","B","UP","SIM","PL"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
